% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enso_costs.R
\name{enso_costs}
\alias{enso_costs}
\title{Natural Disaster Cost Prediction}
\usage{
enso_costs(ONI, risk = 0.811, gdp = 21.9)
}
\arguments{
\item{ONI}{Predicted future ONI value}

\item{risk}{Risk value}

\item{GDP}{Gross Domestic Product of country of interest}
}
\value{
Expected One Year Increase in Natural Disaster Costs Attributed to El Nino (Billions of USD)
}
\description{
This function predicts the natural disaster costs for a specific location based on monthly Ocean Nino Index (ONI) values, risk factor, and GDP
}
\details{
Part of a series of ENSO functions in library(enso)
}
\author{
Austin Richards and Molly Williams
}
