% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enso_extremes.R
\name{enso_extremes}
\alias{enso_extremes}
\title{Impacts of ENSO}
\usage{
enso_extremes(data)
}
\arguments{
\item{data}{Data frame with most recent ENSO values}
}
\value{
Max and min values
}
\description{
This function finds maximum/minimum of historical ENSO data (1950-)
}
\details{
ONI Functions
}
\author{
Austin Richards and Molly Williams
}
