% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enso_plot.R
\name{enso_plot}
\alias{enso_plot}
\title{Impacts of ENSO}
\usage{
enso_plot(data)
}
\arguments{
\item{data}{Data frame with most recent ENSO values}

\item{graph_result}{if TRUE, output a graphical representation of ENSO values over time}
}
\value{
Geom_smooth plot
}
\description{
This functions produces a figure of ONI values over time (1950-)
}
\details{
Part of a series of ENSO functions in library(enso)
}
\author{
Austin Richards and Molly Williams
}
