% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_temp.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Temperature Prediction}
\usage{
x \%!in\% y
}
\arguments{
\item{data}{Data frame with most recent ENSO values}

\item{ONI}{Predicted future ONI value}

\item{month}{Month for which you are predicting temperature}
}
\value{
Temperature prediction (celsius)
}
\description{
This function predicts temperature for a specific location by creating a linear model based on ONI values
}
\details{
ENSO Functions
}
\author{
Austin Richards and Molly Williams
}
