% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_precip.R
\name{predict_precip}
\alias{predict_precip}
\title{Precipitation Prediction}
\usage{
predict_precip(data, ONI = 0)
}
\arguments{
\item{data}{Data frame with ONI, temperature and precipitation values by month for a specific location}

\item{ONI}{Predicted future ONI value}
}
\value{
Precipitation prediction (rainfall in inches)
}
\description{
This function predicts precipitation for a specific location by creating a linear model based on ONI values and historic precipitation values for a specific region
}
\details{
Part of a series of ENSO functions in library(enso)
}
\author{
Austin Richards and Molly Williams
}
