% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_temp.R
\name{predict_temp}
\alias{predict_temp}
\title{Temperature Prediction}
\usage{
predict_temp(data, month, ONI = 0)
}
\arguments{
\item{data}{Data frame with ONI, temperature and precipitation values by month for a specific location}

\item{month}{Month for which you are predicting temperature}

\item{ONI}{Predicted future ONI value}
}
\value{
Temperature prediction (celsius)
}
\description{
This function predicts temperature for a specific location by creating a linear model based on Ocean Nino Index (ONI) values
}
\details{
Part of a series of ENSO functions in library(enso)
}
\author{
Austin Richards and Molly Williams
}
