% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enso_costs.R
\name{enso_costs}
\alias{enso_costs}
\title{Temperature Prediction}
\usage{
enso_costs(ONI, risk = 0.811, gdp = 21.9)
}
\arguments{
\item{ONI}{Predicted future ONI value}

\item{risk}{Risk value}

\item{GDP}{Gross Domestic Product of country of interest}
}
\value{
Expected One Year Increase in Natural Disaster Costs Attributed to El Nino (Billions of USD)
}
\description{
This function predicts temperature for a specific location based on ONI values
}
\details{
ENSO Functions
}
\author{
Austin Richards and Molly Williams
}
